revert breaking changes in 8e9213b, 5225d63
these commits accidentally introduce a breaking
change - if you manually set a field in a model
that is a subclass of another model, that field
gets overwritten (typically with null) on model
creation.

Example: `CatModel` and `DogModel` both extend
`AnimalModel` and set the `family` property to
`'canine'` and `'feline'` respectively.

When either `CatModel` or `DogModel` are created,
the `family` field will be overwritten when
the modeldata (which doesn't have a family value)
is copied to the model.